Updated Eventloop

+++ Configurations/VBS_ZV/EventLoop/Data.h
@@ -17,7 +17,7 @@ public:
   /**
   * @brief Tree variables
   */
-  Double event;
+  int event;
